Properly handle 4xx results
Motivation:

A service might request something and get, for example, a 408 payment
required error. At this moment, aleph will throw an exception and we
will log it. But we don't really want that. We want to short circuit and
return the error that we got.

Modifications:

Add `:throw-exceptions? false` to aleph config.

Result:

We don't get an exception, but the actual request. After that we try to
infer the success based on the response status code.